Add config_clusterHomeServiceMode
When set to 0, ClusterHomeService is in FULL mode, and enabled only when
all core VHAL properties are available. This is same as the current
behavior.
When set to 1, ClusterHomeService is enabled in LIGHT mode, regardless
of whether VHAL properties are available or not. In LIGHT mode, all
service methods corresponding to the core VHAL properties will throw
IllegalStateException, whether the property is available or not.

Bug: 299178284
Bug: 363281818

Test: atest ClusterHalServiceTest ClusterHomeServiceUnitTest \
            ClusterHomeManagerTest

Change-Id: Ia7e16c0ba06bfe348d7440d6832e43e02042a0c7
(cherry picked from commit 8364b91cf9903de949f77e63e31196eb3502ca38)