Initialize PerUserInfo when the user is starting
Currently the PerUserInfo is initialized when the user is unlocked.
If the sender requests a connection and the sender user is still
locked, the connection request will fail because car service needs
to use the sender user's PerUserInfo to get the sender's PackageInfo.
This doesn't make sense. The connection request should depend on
the user state of receiver, but shouldn't depend on the user state
of the sender.

To fix that, this CL initializes the PerUserInfo when the user is
starting, which is earlier than user is unlock. Thus the connection
request can be proceeded even if the sender user is locked.

Fixes: 300676850
Bug: 278136331
Test: atest com.android.car.occupantconnection.CarRemoteDeviceServiceTest
Test: manual test in Control Center
Change-Id: I1cd60d48d64e7e826024ab285a4575e04118105d
(cherry picked from commit 1dcd8e8a379f30fb78d525f7a6cac410863ed769)