Do not remove activity for inactive display.
FixedActivityService can be requested to start an activity while a
display is off. Currently the activity is removed from the list and
can never be started.
Instead, keep the activity in the list and just skip launching the
activity. Also listen to display changes, and later launch the activity
when the display turns on.

Bug: 302223539

Test: atest FixedActivityServiceTest

Change-Id: I29674535b7541878cdb43ababe623660e7da7d6f
(cherry picked from commit 9b02ed3fb11868f8b7cfe5f0c87cf8ebd6154ce9)