Make getProperty backward compatible.
In Android S and T, getProperty might return null if the propertyId
is not supported (property config not found) or might throw exception
if the property Id belongs to user HAL. In android U, we changed the
behavior to throw exception for all unsupported properties.

The behavior for user HAL prop Ids is not consistent with other
unsupported property Ids in S and T, so we update the comment to
mention that client should not use them.

This CL also add logic to maintain backward compatibility for S and T
clients.

Test: CarPropertyManagerUnitTest
Bug: 310251135
Change-Id: Id660852f5663372c82f3f8970ad71bc8d71b199e
(cherry picked from commit e9aafa13dcceafab532ad2c37338dc5c3c08a6aa)