Update car COD to account for format bits so we can be Car Audio again
Our minor class is 0x8 - Car Audio, but that is supposed to set the 5th
bit to account for the format bits, not the 3rd bit. This is causing our
device to be interpreted as an 0x2 - Handsfree Device by remote devices

This change updates the byte in question, as well as the documentation
so its clearer how COD gets applied by spec, in case we ever need to
change it in the future.

Bug: 372049423
Bug: 312759291
Test: make -j; adb logcat | grep -ie "Using class of device"
Test: Boot device, check HCI snoop logs for HCI COD Write packet
Change-Id: I7eb93e2bb3c0aaf96c95cf519f411969f76b19cf
Merged-In: I7eb93e2bb3c0aaf96c95cf519f411969f76b19cf
(cherry picked from commit 15d16f307655c4369aaea10ba8251e8472351124)