Set up DistantDisplay with VirtualDisplay
As DrirverUI is not needed, applied the legacy logic to start
DistantDisplay with two virtual displays on top of physical display.
Newly created VirtualDisplayController and instantiated from the
DistantDisplayActivity as the SurfaceView should be handled under
Activity context. Once a virtual display is created, sends a broadcast
to TaskViewController with displayId of distant display.

Bug: 323443376
Test: manual test
Change-Id: I8e0676e64438bd536bdbcb01f610f371787ff8b2