Initialize PerUserInfo when the user becomes visible
Currently the PerUserInfo is added when a user is starting or unlocked,
and is removed when the user becomes invisible. But when an unlocked
user becomes visible, the PerUserInfo won't be added.

To fix that, this CL addes the PerUserInfo on user visible event.

Fixes: 328154472
Test: atest com.android.car.occupantconnection.CarRemoteDeviceServiceTest
Test: atest android.car.cts.CarRemoteDeviceManagerTest
Change-Id: I19a8e3ac4ee2dd90bfc5a12ebea4a72f70714408
(cherry picked from commit 30b3a1cdd28fb43365b5f130b5b24e33b3c42bd6)