Cancel ChoiceDialog when BugReportActivity is started
Currently, users can open a choice dialog even when a voice recording
dialog is displayed which might be confusing to them.

This CL creates a static variable to track the status of
BugReportActivity and uses it to decide whether a choice dialog
should be shown or not.

Bug: 341848045
Test: Run `adb shell am start -a com.android.car.bugreport.action.START_BUG_REPORT` when a voice recording dialog is shown.
Change-Id: Ibe5d753d69c55ce02990580187879ad7bc3b3650
(cherry picked from commit 0a240ca4e8ea0aefff171992b6214dcfa4145021)