Make test more stable.
This CL adds retry logic inside the test to make the test more
stable. By nature, the test is flaky if the system stuck for
a while (100ms).

This CL also waits for the handler thread to finish after executing
each test to make sure we always create a new thread for each test
case.

Flag: EXEMPT test fix
Test: atest --host FakeVehicleStubUnitTest
Bug: 347196669
Change-Id: Ie9847e644af17d65bce0ef8a393f338a255b3214