Change car audio device configuration schema
This allows for the configurations to be more extendible. The new schema
is as follows:

<deviceConfigurations>
  <deviceConfiguration name="useCoreAudioVolume" value="true"/>
  <deviceConfiguration name="useCoreAudioRouting" value="true"/>
  <deviceConfiguration name="useHalDuckingSignals" value="true"/>
</deviceConfigurations>

Where name is the name of the config, and value is the expected value
which will be parse according to the configuration.
Note: For a repeated configuration name the first encountered value will be
kept.

Bug: 344678175
Flag: android.car.feature.audio_vendor_freeze_improvements
Test: atest CarAudioZonesHelperTest
Change-Id: If7784fdd11936c3804f943e2044112e2ba7613a2