Add a PSI monitor script and sample CUJ execution script.
The PSI monitor script is an on device script that needs to be pushed to
the adb device before executing the script. This script will collect
PSI metrics and dump it into a dump.txt file and wait for CUJ completion
signal. On receiving this signal, this script will monitor PSI metrics
for key changes, such as exceeding/dropping below threshold and
reaching a baseline PSI. When either of these events occur, the PSI
monitor will generate signals to inform any external entity waiting for
the signal.

The run_cuj.sh script contains logic to run sample CUJs (such as
bootup, user switch, app startup, and bootup with app lauches) and
collect PSI measurements. This script serves as an example or a tool to
collect CUJ based PSI measurements.

Test: ./run_cuj.sh -d 90 -g 20 -t 20 --bootup_with_app_launch \
threshold_met;\
./run_cuj.sh -d 90 -g 20 -t 20 --bootup_with_app_launch baseline_met;\
./run_cuj.sh -d 90 -g 20 -a com.android.contacts --app_startup_times\
10;\
./run_cuj.sh -d 90 -g 20 -u 10-11;\
./run_cuj.sh -d 90 -g 20 -t 20 -b
Fixes: 362534488
Flag: EXEMPT test script

Change-Id: I7ff17e1c30a5b3ed450d7c91f3671fd7ca97071f