Fix the STR issue with Guest user
During STR guest was not replaced correctly, A guest was created but
switching to the guest was not working correctly. The CL passes the
request type for initialUserSetter and switch user or set boot user
depending on the context.

TODO (b/374315014): add more automated end to end tests.

Flag: EXEMPT bug fix

Bug: 374315014
Bug: 369634987
Test: atest InitialUserSetterTest
Test: Manual. The change was tested using the lshal. Shell command was used to mock the HAL behavior; STR and on boot behavior was tested.

Change-Id: I6b49ada7c95d8a0dc46aac869f79cf43ce907924