Close the Payload object in car lib.
Close the Payload object so that we will not leak resources. The client
does not care about whether the payload is closed or not since it only
affects the internal shared memory file that is only created during
writeToParcel and only used at readFromParcel. It will not affect
clients calling public methods.

We still require client to call close if sendPayload is not called
although it is technically not necessary since shared memory file
will not be created unless sendPayload is called. However, as a
closable, it is better to close it in case we add more internal
resources in the future.

Flag: EXEMPT bug fix, possible to backport.
Test: Presubmit
Bug: 376532299
Change-Id: I2240adf0d2c7ff1f47bb5942fd21ce4be31fef82