Fix binder death recipient resource mgmt issue.
Fix a bug where if client resets the VhalClient object during
the onBinderDied callback, the thread will block forever. This is
because AidlVhalClient destructor is blocked on onBinderUnlinked
to be called, which is blocked on onBinderDied to return. If
onBinderDied is waiting for the destructor to return, this causes
dead lock.

This CL removes the blocking logic in destructor. In order to make
sure onBinderUnlinked or onBinderDied can be invoked after the
AidlVhalClient is destroyed, we use a cookie object that may live
longer than the AidlVhalClient. The cookie's lifecycyle is managed
by the libinder lib and is not associated with AidlVhalClient
lifecycle. It stores a weak reference to the necessary data structure
managed by AidlVhalClient. If the client goes away before the callback
is invoked, it is still safe to access the cookie but we will do
nothing because the actual object pointed to by the weak pointer
is already gone.

Flag: EXEMPT bug fix
Test: atest AidlVhalClientTest
Bug: 389074926
(cherry-picked from commit: 9e59de81764f560c2573b45f00edf7523486abe3)
Merged-In: Ie372330256c3a1f4d4e033acf053098da7c6954e
Change-Id: Ie372330256c3a1f4d4e033acf053098da7c6954e