Add activity launch test to KitchenSink
This feature adds a new menu option "activity launch" to the
KitchenSink app. This option opens a fragment with three buttons
to launch a test activity in the same task, a new task, or a new
task with the launch adjacent flag. The test activity also contains
the same options to launch a second test activity.

Test: Manual verification on an Android Automotive OS device.
Bug: 433539393
Change-Id: I03efbe087765614381833cc9ffd30885352997ff
Merged-In: I03efbe087765614381833cc9ffd30885352997ff